Move Route out of Solution
